file_attributes - move FileTypes to its own file
