+++ src/model/mod.rs
@@ -1,5 +1,6 @@
 pub mod api;
 pub mod error;
+pub mod file_types;
 pub mod guard;
 pub mod repository;
 pub mod request;
